fix invalid type for facilitiesUpdated in TFM schema

+++ trade-finance-manager-api/src/graphql/schemas/index.js
@@ -342,7 +342,7 @@ type DealSnapshot {
   isFinanceIncreasing: Boolean
   eligibility: DealEligibility
   dealFiles: DealFiles
-  facilitiesUpdated: Int
+  facilitiesUpdated: Float
 }
 
 type Deal {
